Change Utils Dir Strucutre
