+++ tt-media-server/cpp_server/benchmarks/tokenizer_benchmark.cpp
@@ -9,7 +9,7 @@
 #include <vector>
 
 #include "config/settings.hpp"
-#include "utils/tokenizer.hpp"
+#include "utils/tokenizers/tokenizer.hpp"
 
 using namespace tt::utils;
