+++ tt-media-server/cpp_server/src/utils/tokenizers/deepseek_tokenizer.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
 
-#include "utils/deepseek_tokenizer.hpp"
+#include "utils/tokenizers/deepseek_tokenizer.hpp"
 
 #include <sstream>
